+++ frontend/src/pages/MainPage/MainPage.tsx
@@ -8,7 +8,7 @@ import { useGetCardList } from '@/hooks/Queries/useClub';
 import Banner from '@/pages/MainPage/components/Banner/Banner';
 import CategoryButtonList from '@/pages/MainPage/components/CategoryButtonList/CategoryButtonList';
 import ClubCard from '@/pages/MainPage/components/ClubCard/ClubCard';
-import Filter from '@/pages/MainPage/components/Filter/Filter';
+import Filter from '@/components/common/Filter/Filter';
 import Popup from '@/pages/MainPage/components/Popup/Popup';
 import { useSelectedCategory } from '@/store/useCategoryStore';
 import { useSearchIsSearching, useSearchKeyword } from '@/store/useSearchStore';
